Replace 'model' option with 'no-model' switch
Using `--model=false` doesn't work, and reading the text description
that would appear to be the way to toggle this switch, however, `false`
gets treated as a string. Using `--model=0` does work but isn't obvious.
This replaces `--model` option with `--no-model` switch instead. The
default is to create the model (as it currently does) but to disable
the creation of the model only `--no-model` needs to be supplied.