Don't duplicate id/created_at/updated_at fields if they are passed in
The current migrations assume that everyone wants the default id and
timestamp definitions added, but this isn't always the case. It now
checks to see if any schema definitions have been passed in for id,
created_at and updated_at and if they have then the default fields
aren't added.